Grimoire, magic switching and score resetting
Adds widget functionality to right hand, temporarily replacing ability to cast magic. Allows the player to select from a list of buttons a magic type, which will be allocated to their controller and change which magic they will cast. Adds a temporary button that resets the player's score when grabbed.